Create valuestack
This lives separately to the heap, and contains pointers that would
normally be on the regular C stack.

This stack holds all the live references, including the symbol table
and the root environment.

Any Pointers in C code should now be replaced by stack indexes, which
will serve two purposes:
1. The indexes are invariant and stay valid after a GC phase.
2. The indexes serve as a live reference, meaning that stack objects
   won't disappear after a GC phase.