Active heap objects point to themselves
This is a pre-requisite for two-finger GC.

Previously we were using a relocated value of nil to indicate this
object is on the active heap. This is fine, but for two-finger GC we
need to be able to identify a heap object without having a pointer.

This change allows us to identify objects without knowing their type,
which has the side-benefit of being able to verify the validity of a
pointer by checking that pointer->relocated == pointer.