vector_set takes index not pointer for vector
This solves a nasty and subtle bug in force-GC mode where seemingly-safe
stack dereferences get clobbered.

eg. This looks right:

    vector_set(GET(valueIndex), 1, map_analyse(argsIndex));

But it appears it was being evaluated left-to-right, and the pointer
gotten from GET(valueIndex) was no longer valid after map_analyse(...).

Passing in an index for the first one solves the problem. pair_set
should use this too.